+++ packages/chakra-ui-core/src/CAspectRatioBox/tests/CAspectRatioBox.test.js
@@ -6,7 +6,9 @@ const renderComponent = (props) => {
     components: { CAspectRatioBox, CBox },
     template: `
     <CAspectRatioBox maxW="400px" data-testid="aspectRatioBox" ${inlineAttrs}>
-      <CBox as="img" src="https://bit.ly/naruto-sage" alt="naruto" objectFit="cover" data-testid="image" />
+      <CBox data-testid="child">
+        <CBox as="img" src="https://bit.ly/naruto-sage" alt="naruto" objectFit="cover" h="100%" w="100%" data-testid="image" />
+      </CBox>
     </CAspectRatioBox>
     `,
     ...props
@@ -20,37 +22,25 @@ it('should render correctly', () => {
   expect(asFragment()).toMatchSnapshot()
 
   const [emotionClassName] = [...screen.getByTestId('aspectRatioBox').classList]
-  const pseudoStyles = getElementStyles(`.${emotionClassName}:before`)
+  const pseudoStyles = getElementStyles(`.${emotionClassName}::before`)
 
-  expect(pseudoStyles).toContain(`
+  expect(pseudoStyles).toContain(
+    `
     padding-bottom: 100%
-  `.trim())
+  `.trim()
+  )
 })
 
 it('should have correct styles', () => {
   const inlineAttrs = ':ratio="2"'
   renderComponent({ inlineAttrs })
-  const image = screen.getByTestId('image')
-  const aspectRatioBox = screen.getByTestId('aspectRatioBox')
 
-  const [emotionClassName] = [...aspectRatioBox.classList] // second className has the pseudo styles
-  const pseudoStyles = getElementStyles(`.${emotionClassName}:before`)
+  const [emotionClassName] = [...screen.getByTestId('aspectRatioBox').classList]
+  const pseudoStyles = getElementStyles(`.${emotionClassName}::before`)
 
-  expect(pseudoStyles).toContain(`
+  expect(pseudoStyles).toContain(
+    `
     padding-bottom: 50%
-  `.trim())
-
-  expect(aspectRatioBox).toHaveStyle(`
-    max-width: 400px;
-    position: relative;
-  `)
-
-  expect(image).toHaveStyle(`
-    object-fit: cover;
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0px;
-    left: 0px;
-  `)
+  `.trim()
+  )
 })