+++ packages/chakra-ui-core/src/CImage/CImage.js
@@ -25,6 +25,7 @@ const CImage = {
   mixins: [createStyledAttrsMixin('CImage')],
   props: {
     src: String,
+    srcset: String,
     fallbackSrc: String,
     ignoreFalback: Boolean,
     htmlWidth: String,
@@ -48,13 +49,18 @@ const CImage = {
   created () {
     // Should only invoke window.Image in the browser.
     if (process.browser) {
-      this.loadImage(this.src)
+      this.loadImage(this.src, this.srcset)
     }
   },
   methods: {
-    loadImage (src) {
+    loadImage (src, srcset) {
       const image = new window.Image()
-      image.src = src
+      if (srcset) {
+        image.srcset = srcset
+      }
+      if (src) {
+        image.src = src
+      }
 
       image.onload = (event) => {
         this.hasLoaded = true
@@ -70,9 +76,9 @@ const CImage = {
   render (h) {
     let imageProps
     if (this.ignoreFallback) {
-      imageProps = { src: this.src }
+      imageProps = { src: this.src, srcset: this.srcset }
     } else {
-      imageProps = { src: this.hasLoaded ? this.src : this.fallbackSrc }
+      imageProps = { src: this.hasLoaded ? this.src : this.fallbackSrc, srcset: this.srcset }
     }
     return h(CNoSsr, [
       h('img', {