+++ packages/chakra-ui-theme/src/theme/shadows.js
@@ -9,7 +9,7 @@ const shadows = {
     '0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)',
   '2xl': '0 25px 50px -12px rgba(0, 0, 0, 0.25)',
   outline: '0 0 0 3px rgba(66, 153, 225, 0.6)',
-  inner: 'inset 0 2px 4px 0 rgba(0,0,0,0.06)',
+  inner: 'inset 0 2px 4px 0 rgba( 0, 0, 0, 0.06)',
   none: 'none'
 }