+++ packages/chakra-ui-core/CHANGELOG.md
@@ -1,14 +1,155 @@
 # Change Log
 
+## 0.10.1
+
+### Minor Changes
+
+- [`7d1bdf30`](https://github.com/chakra-ui/chakra-ui-vue/commit/7d1bdf30b4b1f5fab48f2c23bcdea4038ed7cf5f) [#474](https://github.com/chakra-ui/chakra-ui-vue/pull/474) Thanks [@simpletrontdip](https://github.com/simpletrontdip)! - Backport v1 Theme API to v0.x
+- [`2bb4540`](https://github.com/chakra-ui/chakra-ui-vue/pull/473/commits/2bb4540dbd5abeaed9dfbf5108a0f345a9e25bf6)[#473](https://github.com/chakra-ui/chakra-ui-vue/pull/473) Thanks [@Pwuts](https://github.com/Pwuts) - feat(types): improve & add type declarations
+
+### Patch Changes
+
+- [`4429524b`](https://github.com/chakra-ui/chakra-ui-vue/commit/4429524b01a00a31a272d258ad45d6edee439f6b) [#476](https://github.com/chakra-ui/chakra-ui-vue/pull/476) Thanks [@codebender828](https://github.com/codebender828)! - version release
+
+## 0.10.0
+
+### Minor Changes
+
+- [`7d1bdf30`](https://github.com/chakra-ui/chakra-ui-vue/commit/7d1bdf30b4b1f5fab48f2c23bcdea4038ed7cf5f) [#474](https://github.com/chakra-ui/chakra-ui-vue/pull/474) Thanks [@simpletrontdip](https://github.com/simpletrontdip)! - Backport v1 Theme API to v0.x
+- [`2bb4540`](https://github.com/chakra-ui/chakra-ui-vue/pull/473/commits/2bb4540dbd5abeaed9dfbf5108a0f345a9e25bf6)[#473](https://github.com/chakra-ui/chakra-ui-vue/pull/473) Thanks [@Pwuts](https://github.com/Pwuts) - feat(types): improve & add type declarations
+
+### Patch Changes
+
+- [`d1e97d75`](https://github.com/chakra-ui/chakra-ui-vue/commit/d1e97d75b82613ee0cad97e1136db1fd9447f6da) [#468](https://github.com/chakra-ui/chakra-ui-vue/pull/468) Thanks [@abelnnieva](https://github.com/abelnnieva)! - fix(menu): update MenuOption defaultValue prop to inner value
+
+## 0.9.4
+
+### Patch Changes
+
+- fix: syncs core and nuxt packages
+
+## 0.9.3
+
+### Patch Changes
+
+- [`46a353a`](https://github.com/chakra-ui/chakra-ui-vue/commit/46a353a4348ea5829834f0ad479026ca647002b1) [#459](https://github.com/chakra-ui/chakra-ui-vue/pull/459) Thanks [@abelnnieva](https://github.com/abelnnieva)! - fix: add an unstyled variant to the c-button validators
+
+* [`826eed2`](https://github.com/chakra-ui/chakra-ui-vue/commit/826eed2f2b04b545e0379381c79ab8cd8b2c54eb) [#460](https://github.com/chakra-ui/chakra-ui-vue/pull/460) Thanks [@simpletrontdip](https://github.com/simpletrontdip)! - Pass all listeners to render function in CFlex, Stack
+
+## 0.9.2
+
+### Patch Changes
+
+- [`2627e9f`](https://github.com/chakra-ui/chakra-ui-vue/commit/2627e9f833ec62d4bc35703bcb3bb1293d20c391) [#455](https://github.com/chakra-ui/chakra-ui-vue/pull/455) Thanks [@codebender828](https://github.com/codebender828)! -
+
+## 0.9.1
+
+### Patch Changes
+
+- [`f124d5c`](https://github.com/chakra-ui/chakra-ui-vue/commit/f124d5c48d023882a7bd94c96e46a5723bd9a85d) [#446](https://github.com/chakra-ui/chakra-ui-vue/pull/446) Thanks [@codebender828](https://github.com/codebender828)! - Fixes CImage src attribute bug
+
+## 0.9.0
+
+### Minor Changes
+
+- [`5552440`](https://github.com/chakra-ui/chakra-ui-vue/commit/5552440eb8e0f1fb283c7dc263888bf09dcfa7b9) [#438](https://github.com/chakra-ui/chakra-ui-vue/pull/438) Thanks [@codebender828](https://github.com/codebender828)! - - This changeset adds a srcset support to the CImage component.
+  - Fixes merging of Popper.js modifiers
+  - Adds CGridItems components to the core 👏🏾
+
+## 0.8.1
+
+### Patch Changes
+
+- [`e640030`](https://github.com/chakra-ui/chakra-ui-vue/commit/e640030d277966a384261d1a90a1820335c4c007) [#418](https://github.com/chakra-ui/chakra-ui-vue/pull/418) Thanks [@codebender828](https://github.com/codebender828)! - Fix for close button to change color
+
+## 0.8.0
+
+### Minor Changes
+
+- [`8af1c07`](https://github.com/chakra-ui/chakra-ui-vue/commit/8af1c07404d2cfa94bb338e4268c5e0869a50776) Thanks [@codebender828](https://github.com/codebender828)! - Adds type definitions, aspect-ratio-box fixes, switch fixes and collapse fixes
+- feat: add type definition of component #380 by @odanado
+- feat: CCollapse now renders with collapsed height if initially closed by @codebender828 . closes #269
+- fix: Set ccollapse height to auto on enter complete, unless finalHeight wa... #384 by @IHIutch
+- fix: Add position: relative to switch label #396 by @IHIutch
+- fix: Deeply nested aspect-ratio by @IHIutch
+
+## 0.7.4
+
+### Patch Changes
+
+- [`951906a`](https://github.com/chakra-ui/chakra-ui-vue/commit/951906a61deabdcf1d3f04b7f8f9fdeaad1f3841) [#375](https://github.com/chakra-ui/chakra-ui-vue/pull/375) Thanks [@codebender828](https://github.com/codebender828)! - Added support for Feather Icons to Nuxt.js
+
+* fix(nuxt): nuxt module icon parsing
+
+## 0.7.3
+
+## 0.7.2
+
+### Patch Changes
+
+- 5db9572: Fix build for nuxt.js
+- 5db9572: Export utils
+
+## 0.7.2-next.0
+
+### Patch Changes
+
+- Fix build for nuxt.js
+- Export utils
+
+## 0.7.1
+
+### Patch Changes
+
+- Improve publishing workflow
+
+## 0.7.0
+
+### Minor Changes
+
+- 8715f8c: Upgraded `emotion` dependencies to namespaced `@emotion/css` packages
+
+### Patch Changes
+
+- 6b21ca5: Upgraded to @nuxtjs/emotion@^0.1.0
+- 98ccc9f: Added support for feather-icons to Chakr aUI Vue
+- eb47ac5: Removes UMD builds for Chakra UI Vue
+
+## 0.7.0-next.3
+
+### Patch Changes
+
+- Removes UMD builds for Chakra UI Vue
+
+## 0.7.0-next.2
+
+### Patch Changes
+
+- Added support for feather-icons to Chakr aUI Vue
+
+## 0.7.0-next.1
+
+### Patch Changes
+
+- Upgraded to @nuxtjs/emotion@^0.1.0
+
+## 0.7.0-next.0
+
+### Minor Changes
+
+- Upgraded `emotion` dependencies to namespaced `@emotion/css` packages
+
 ## 0.6.6
 
 ### Patch Changes
 
 **🌟 New features**
+
 - feat: include all known css properties #319
 - feat(CStack): allow stack html elements #324
 
 **🛠 Bug fixes**
+
 - fix(creset): include font-family to global styles #323
 - fix(inputgroup): default inputgroup z-index to containing element zindex #325
 - Documentation fixes #329, #328, #321